refactor: move BatesNumberConfig to src/Config/ for discoverability
