+++ src/BatesSequence.cs
@@ -1,14 +1,7 @@
 using System.Globalization;
+using Zipper.Config;
 
 namespace Zipper;
-
-public record BatesNumberConfig
-{
-    public string Prefix { get; init; } = "DOC";
-    public long Start { get; init; } = 1;
-    public int Digits { get; init; } = 8;
-    public long Increment { get; init; } = 1;
-}
 
 public record BatesNumber(string Prefix, string FormattedNumber)
 {
