+++ src/Cli/Validation/CrossCuttingValidator.cs
@@ -1,3 +1,4 @@
+using Zipper.Config;
 using Zipper.Profiles;
 
 namespace Zipper.Cli.Validation;
